+++ mobile/android/android-components/components/lib/state/src/main/java/mozilla/components/lib/state/Store.kt
@@ -4,17 +4,13 @@
 
 package mozilla.components.lib.state
 
-import android.os.Handler
-import android.os.Looper
 import androidx.annotation.CheckResult
 import androidx.annotation.VisibleForTesting
-import kotlinx.coroutines.CoroutineExceptionHandler
 import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.asCoroutineDispatcher
-import kotlinx.coroutines.cancel
 import kotlinx.coroutines.launch
+import mozilla.components.lib.state.internal.DefaultStoreDispatcher
 import mozilla.components.lib.state.internal.ReducerChainBuilder
-import mozilla.components.lib.state.internal.StoreThreadFactory
+import mozilla.components.lib.state.internal.StoreDispatcher
 import java.lang.ref.WeakReference
 import java.util.Collections
 import java.util.concurrent.ConcurrentHashMap
@@ -29,34 +25,38 @@ import java.util.concurrent.Executors
  * @param initialState The initial state until a dispatched [Action] creates a new state.
  * @param reducer A function that gets the current [State] and [Action] passed in and will return a new [State].
  * @param middleware Optional list of [Middleware] sitting between the [Store] and the [Reducer].
- * @param threadNamePrefix Optional prefix with which to name threads for the [Store]. If not provided,
- * the naming scheme will be deferred to [Executors.defaultThreadFactory]
  */
-open class Store<S : State, A : Action>(
+open class Store<S : State, A : Action> internal constructor(
     initialState: S,
     reducer: Reducer<S, A>,
-    middleware: List<Middleware<S, A>> = emptyList(),
-    threadNamePrefix: String? = null,
+    middleware: List<Middleware<S, A>>,
+    dispatcher: StoreDispatcher,
 ) {
-    private val threadFactory = StoreThreadFactory(threadNamePrefix)
-    private val dispatcher = Executors.newSingleThreadExecutor(threadFactory).asCoroutineDispatcher()
-    private val reducerChainBuilder = ReducerChainBuilder(threadFactory, reducer, middleware)
-    private val scope = CoroutineScope(dispatcher)
+
+    /**
+     * @param initialState The initial state until a dispatched [Action] creates a new state.
+     * @param reducer A function that gets the current [State] and [Action] passed in and will return a new [State].
+     * @param middleware Optional list of [Middleware] sitting between the [Store] and the [Reducer].
+     * @param threadNamePrefix Optional prefix with which to name threads for the [Store]. If not provided,
+     * the naming scheme will be deferred to [Executors.defaultThreadFactory]
+     */
+    constructor(
+        initialState: S,
+        reducer: Reducer<S, A>,
+        middleware: List<Middleware<S, A>> = emptyList(),
+        threadNamePrefix: String? = null,
+    ) : this(
+        initialState = initialState,
+        reducer = reducer,
+        middleware = middleware,
+        dispatcher = DefaultStoreDispatcher(threadNamePrefix),
+    )
+
+    private val reducerChainBuilder = ReducerChainBuilder(dispatcher, reducer, middleware)
+    private val scope = CoroutineScope(dispatcher.coroutineContext)
 
     @VisibleForTesting
     internal val subscriptions = Collections.newSetFromMap(ConcurrentHashMap<Subscription<S, A>, Boolean>())
-    private val exceptionHandler = CoroutineExceptionHandler { _, throwable ->
-        // We want exceptions in the reducer to crash the app and not get silently ignored. Therefore we rethrow the
-        // exception on the main thread.
-        Handler(Looper.getMainLooper()).postAtFrontOfQueue {
-            throw StoreException("Exception while reducing state", throwable)
-        }
-
-        // Once an exception happened we do not want to accept any further actions. So let's cancel the scope which
-        // will cancel all jobs and not accept any new ones.
-        scope.cancel()
-    }
-    private val dispatcherWithExceptionHandler = dispatcher + exceptionHandler
 
     @Volatile private var currentState = initialState
 
@@ -91,7 +91,7 @@ open class Store<S : State, A : Action>(
     /**
      * Dispatch an [Action] to the store in order to trigger a [State] change.
      */
-    fun dispatch(action: A) = scope.launch(dispatcherWithExceptionHandler) {
+    fun dispatch(action: A) = scope.launch {
         synchronized(this@Store) {
             reducerChainBuilder.get(this@Store).invoke(action)
         }
@@ -179,9 +179,3 @@ open class Store<S : State, A : Action>(
         }
     }
 }
-
-/**
- * Exception for otherwise unhandled errors caught while reducing state or
- * while managing/notifying observers.
- */
-class StoreException(msg: String, val e: Throwable? = null) : Exception(msg, e)